+++ CHANGELOG.md
@@ -4,7 +4,9 @@ CHANGELOG
 0.2.0
 -----
 
-* made the client compatible with version 0.2 of the `php-xapi/model` package
+* made the client compatible with version 0.5 of the `php-xapi/model` package
+
+* made the client compatible with version 0.3 of the `php-xapi/serializer` package
 
 0.1.0
 -----